+++ src/senaite/core/api/geo.py
@@ -18,10 +18,13 @@
 # Copyright 2018-2024 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-from six import string_types
-
+import math
 import pycountry
+
+from bika.lims.api import is_floatable
+from bika.lims.api import to_float
 from bika.lims.api import to_utf8
+from six import string_types
 
 _marker = object()
 
@@ -199,3 +202,139 @@ def get_country_or_subdivision(thing, default=_marker):
         raise ValueError("Could not find a record for '{}'".format(
             thing.lower()))
     return default
+
+
+def to_dms(degrees, precision=4, default=_marker):
+    """Converts a geographical coordinate in decimal degrees to a dict with
+    degrees, minutes and seconds as the keys
+
+    :param degrees: coordinate in decimal degrees
+    :type degrees: string,int,float
+    :param precision: number of decimals for seconds
+    :type precision: int
+    :return: a dict with degrees, minutes and seconds as keys
+    """
+    if not is_floatable(degrees):
+        if default is _marker:
+            raise ValueError("Expected decimal degrees to be a floatable, but "
+                             "got %r" % degrees)
+        return default
+
+    # calculate the DMS
+    decimal_degrees = to_float(degrees)
+    degrees = math.trunc(decimal_degrees)
+    minutes = math.trunc((decimal_degrees - degrees) * 60)
+    seconds = decimal_degrees * 3600 % 60
+
+    # check precision type
+    if not isinstance(precision, int):
+        raise TypeError("Expected precision to be an `int`, but got %r"
+                        % type(precision))
+
+    # apply the precision to seconds
+    template = "{:.%df}" % precision
+    seconds = template.format(seconds)
+
+    return {
+        "degrees": degrees,
+        "minutes": minutes,
+        "seconds": to_float(seconds),
+    }
+
+
+def to_latitude_dms(degrees, precision=4, default=_marker):
+    """Converts a geographical latitude in decimal degrees to a dict with
+    degrees, minutes, seconds and bearing as the keys
+
+    :param degrees: latitude in decimal degrees
+    :type degrees: string,int,float
+    :param precision: number of decimals for seconds
+    :type precision: int
+    :return: a dict with degrees, minutes, seconds and bearing as keys
+    """
+    if not is_floatable(degrees):
+        if default is _marker:
+            raise ValueError("Expected decimal degrees to be a floatable, but "
+                             "got %r" % degrees)
+        return default
+
+    # check latitude is in range
+    latitude = to_float(degrees)
+    if abs(latitude) > 90:
+        if default is _marker:
+            raise ValueError("Latitude must be within -90 and 90 degrees")
+        return default
+
+    # calculate the DMS
+    dms = to_dms(abs(latitude), precision=precision)
+    dms["bearing"] = "N" if latitude >= 0 else "S"
+    return dms
+
+
+def to_longitude_dms(degrees, precision=4, default=_marker):
+    """Converts a geographical longitude in decimal degrees to a dict with
+    degrees, minutes, seconds and bearing as the keys
+
+    :param degrees: longitude in decimal degrees
+    :type degrees: string,int,float
+    :param precision: number of decimals for seconds
+    :type precision: int
+    :return: a dict with degrees, minutes, seconds and bearing as keys
+    """
+    if not is_floatable(degrees):
+        if default is _marker:
+            raise ValueError("Expected decimal degrees to be a floatable, but "
+                             "got %r" % degrees)
+        return default
+
+    # check longitude is in range
+    longitude = to_float(degrees)
+    if abs(longitude) > 180:
+        if default is _marker:
+            raise ValueError("Longitude must be within -180 and 180 degrees")
+        return default
+
+    # calculate the DMS
+    dms = to_dms(abs(longitude), precision=precision)
+    dms["bearing"] = "E" if longitude >= 0 else "W"
+    return dms
+
+
+def to_decimal_degrees(dms, precision=7, default=_marker):
+    """Converts a geographical coordinate in DMS format to decimal degrees
+
+    :param dms: coordinate in DMS
+    :type dms: dict
+    :param precision: number of decimals for decimal degrees
+    :type precision: int
+    :return: a float representing a geographical coordinate in decimal degrees
+    """
+    if not isinstance(dms, dict):
+        if default is _marker:
+            raise TypeError("Expected dms to be a dict, but got %r" % dms)
+        return default
+
+    # get the degrees, minutes and seconds
+    degrees = to_float(dms.get("degrees"), default=0)
+    minutes = to_float(dms.get("minutes"), default=0)
+    seconds = to_float(dms.get("seconds"), default=0)
+
+    # calculate the decimal degrees
+    decimal_degrees = abs(degrees + (minutes / 60) + (seconds / 3600))
+
+    # Use +/- to express N/S, W/E
+    bearing = dms.get("bearing")
+    if bearing and bearing in "SW":
+        decimal_degrees = -decimal_degrees
+
+    # check precision type
+    if not isinstance(precision, int):
+        raise TypeError("Expected precision to be an `int`, but got %r"
+                        % type(precision))
+
+    # apply the precision
+    template = "{:.%df}" % precision
+    decimal_degrees = template.format(decimal_degrees)
+
+    # return the float value
+    return to_float(decimal_degrees)