+++ src/senaite/core/upgrade/v02_06_000.py
@@ -45,12 +45,12 @@ from senaite.core.setuphandlers import add_senaite_setup_items
 from senaite.core.setuphandlers import setup_core_catalogs
 from senaite.core.setuphandlers import setup_other_catalogs
 from senaite.core.upgrade import upgradestep
-from senaite.core.upgrade.utils import UpgradeUtils
 from senaite.core.upgrade.utils import copy_snapshots
 from senaite.core.upgrade.utils import del_metadata
 from senaite.core.upgrade.utils import delete_object
 from senaite.core.upgrade.utils import permanently_allow_type_for
 from senaite.core.upgrade.utils import uncatalog_object
+from senaite.core.upgrade.utils import UpgradeUtils
 from senaite.core.workflow import ANALYSIS_WORKFLOW
 from senaite.core.workflow import LABCONTACT_WORKFLOW
 from zope.component import getMultiAdapter
@@ -83,6 +83,8 @@ REMOVE_AT_TYPES = [
     "StorageLocations",
     "InstrumentType",
     "InstrumentTypes",
+    "SamplingDeviation",
+    "SamplingDeviations",
     "AnalysisCategory",
     "AnalysisCategories",
 ]
@@ -1185,6 +1187,54 @@ def migrate_storagelocations_to_dx(tool):
 
 
 @upgradestep(product, version)
+def migrate_samplingdeviations_to_dx(tool):
+    """Converts existing sampling deviations to Dexterity
+    """
+    logger.info("Convert Sampling Deviations to Dexterity ...")
+
+    # ensure old AT types are flushed first
+    remove_at_portal_types(tool)
+
+    # run required import steps
+    tool.runImportStepFromProfile(profile, "typeinfo")
+    tool.runImportStepFromProfile(profile, "workflow")
+
+    # get the old container
+    origin = api.get_setup().get("bika_samplingdeviations")
+    if not origin:
+        # old container is already gone
+        return
+
+    # get the destination container
+    destination = get_setup_folder("samplingdeviations")
+
+    # un-catalog the old container
+    uncatalog_object(origin)
+
+    # Mapping from schema field name to a tuple of
+    # (accessor, target field name, default value)
+    schema_mapping = {
+        "title": ("Title", "title", ""),
+        "description": ("Description", "description", ""),
+    }
+
+    # migrate the contents from the old AT container to the new one
+    migrate_to_dx("SamplingDeviation",
+                  origin, destination, schema_mapping)
+
+    # copy snapshots for the container
+    copy_snapshots(origin, destination)
+
+    # remove old AT folder
+    if len(origin) == 0:
+        delete_object(origin)
+    else:
+        logger.warn("Cannot remove {}. Is not empty".format(origin))
+
+    logger.info("Convert Sampling Deviations to Dexterity [DONE]")
+
+
+@upgradestep(product, version)
 def migrate_analysiscategories_to_dx(tool):
     """Converts existing analysis categories to Dexterity
     """
@@ -1578,3 +1628,37 @@ def remove_sampler_fullname(tool):
     del_metadata(SAMPLE_CATALOG, "getSamplerFullName")
 
     logger.info("Removing getSamplerFullName from catalogs [DONE]")
+
+
+def migrate_samplepoints_coordinates(tool):
+    """Migrates the values from SamplePoint coordinate fields to GPSCoordinates
+    """
+    logger.info("Migrating coordinates from SamplePoint ...")
+    cat = api.get_tool(SETUP_CATALOG)
+    brains = cat(portal_type="SamplePoint")
+    total = len(brains)
+    for num, brain in enumerate(brains):
+        if num and num % 100 == 0:
+            logger.info("Processed objects: {}/{}".format(num, total))
+
+        obj = api.get_object(brain)
+
+        # get the value from attributes (once DX fields)
+        latitude = getattr(obj, "latitude", None)
+        longitude = getattr(obj, "longitude", None)
+        if not all([latitude, longitude]):
+            # migrated already
+            continue
+
+        # set the value to the new field Location
+        obj.setLatitude(latitude)
+        obj.setLongitude(longitude)
+
+        # remove the attributes for latitude and longitude
+        delattr(obj, "latitude")
+        delattr(obj, "longitude")
+
+        obj.reindexObject()
+        obj._p_deactivate() # noqa
+
+    logger.info("Migrating coordinates from SamplePoint [DONE]")